Move FrameTime to godex module
